requirements.txt
Flask==2.3.3
nltk==3.8.1
sumy==0.11.0
rake-nltk==1.0.6
python-docx==1.1.0
PyMuPDF==1.23.20
wordcloud==1.9.3
matplotlib==3.8.4
gunicorn